IDE: use the right method to read ATAPI drives
This change actually fixes issue #26. The kernel is able to read the primary volume descriptor
and returns the same values for the volume identifier,  volume size, path
table size, path table starting lba and the rootdir starting lba.

While the ATAPI drive can be read, there is an issue where the kernel is
unable to parse the configuration file.
This may be due to an issue with Vireo not handling systems
with memory between 509-564 MB well (data is not correctly copied by memcpy).