ps2keyb: semi-working driver
This driver was intended for scancode set 2, which appears to be non-working on Qemu and Vbox, while it should be the default. Either that, or most sources confuse the scancode sets for each other. While it could be that I did something wrong, I tried doing exactly the same thing as some other drivers, but it just didn't work.

Anyways, continuing with scancode set 1 (according to wiki.osdev.org) everything almost works.

Still need to do some clean-up.