CP: search entire keyboard buffer for characters
Before, it would only read the very first keycode everytime. This is because one character was given when it was found, when reading the buffer the next time, there would be (most likely) a new key at the first index, since the keyboard buffer would have seen the keybuffer as empty. This may cause problems when CP is somehow slower than the user.

The keyboard driver deems a buffer empty when the very first index of the keybuffer is 0.