iso9660: fix driver finding shorter filenames
The driver used to check whether the length of the identifier was larger than the filename length and would then check if the filenames were the same for filename_length characters.

The strlen of the identifier will be larger than the filename_length in most cases, because it is not null termintated.

Because the identifier was checked for filename_length characters, the driver would always find `breaker.elf` (identifier) while looking for filename `break`.

Now, the driver first checks if the identifier actually has the same length as the filename, and only then will the identifier and filename be compared for filename_length characters.