fat32: improved fat_filename_fatcompat()
Sometimes when there was no file extension in a filename (see examples below),
fat_filename_fatcompat() would fail to add spaces leading to a shorter filename.

A shorter filename meant that the driver could not check for filenames
correctly and was unable to find files or directories. The missing spaces
would not happen in all filenames. For example, it would not happen
with a filename 'FDOS', while it would happen for 'PACKAGES'.

This newer version of fat_filename_fatcompat() does not have this problem
anymore.

Examples of filenames with no extension:
- Current directory ('.')
- Parent directory ('..')
- Files with no extension ('FDOS'  and 'FILE' )