copy: add note about file links in copied dirs
When copying directories, the entire directory structure is copied one-to-one. This means that all cluster references are  the same in the copied directory compared to the original directory. This program does not copy all files within a directory. Therefore, a copied directory links to the exact same files as contained in the original directory.